+++ tests/test_info.py
@@ -89,6 +89,7 @@ def test_fund():
     assert hs300.round_label == 1
     assert hs300.name == "景顺长城沪深300指数增强"  ## "景顺长城沪深300增强", 蜜汁改名。。。
     assert hs300.fenhongdate[1] == pd.Timestamp("2017-08-15")
+    assert hs300.get_holdings(2019, 4).iloc[0]["name"] == "中国平安"
     assert (
         float(hs300.special[hs300.special["date"] == "2017-08-04"]["comment"]) == 0.19
     )